requirements.txt
fastapi
uvicorn
redis
pydantic
sqlalchemy
pymysql
python-jose
httpx 